pyproject.toml
[tool.ty.environment]
python = "./.venv/bin/python3"

[project]
name = "midterm3"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "ipython>=9.7.0",
    "matplotlib>=3.10.8",
    "numpy>=2.3.5",
    "pillow>=12.0.0",
    "pybullet>=3.2.7",
    "pyqt6>=6.10.1",
]
